Refactor: Organize variables and constants in upgrades.js
- Group game constants at the top of the file
- Consolidate upgrade system variables
- Move automation arrays (autoplanters and autoHarvesters) to the top
- Reposition the upgrades array definition
- Remove redundant comments and consolidate related variables
- Improve overall code structure and readability

Prompt used: "Please review the following code and reorganize the variables and constants at the top of the file. Group them logically, starting with game constants, followed by system variables, and then any large data structures or arrays. Add clear, concise comments to separate different sections. Ensure that the reorganization doesn't affect the functionality of the code. After reorganizing, provide a brief explanation of the changes made and write a commit title and description for this refactoring."